mountains/pyproject.toml
[tool.poetry]
name = "mountains"
version = "0.1.0"
description = ""
authors = ["Mark Waddoups <mwaddoups@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^3.2"
django-rest-framework = "^0.1.0"
django-cors-headers = "^3.10.1"
drfpasswordless = "^1.5.7"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
